refactor:extract wmts utils
